% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificialLifeR-package.R
\docType{package}
\name{artificialLifeR-package}
\alias{artificialLifeR}
\alias{artificialLifeR-package}
\title{artificialLifeR: Educational simulations for artificial life}
\description{
\code{artificialLifeR} provides simplified simulations for teaching artificial-life
concepts, including agents, resources, reproduction, mutation, selection,
population dynamics, and life-like complexity summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NoushinN/artificialLifeR}
  \item \url{https://noushinn.github.io/artificialLifeR/}
  \item Report bugs at \url{https://github.com/NoushinN/artificialLifeR/issues}
}

}
\author{
\strong{Maintainer}: Noushin Nabavi \email{12345678+NoushinN@users.noreply.github.com}

Authors:
\itemize{
  \item Noushin Nabavi \email{12345678+NoushinN@users.noreply.github.com}
}

}
\keyword{internal}
