% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alife_sim.R
\name{plot_alife_sim}
\alias{plot_alife_sim}
\title{Plot artificial-life simulation outputs}
\usage{
plot_alife_sim(data, x, y, value = NULL, type = c("line", "point", "raster"))
}
\arguments{
\item{data}{Data frame to plot.}

\item{x}{Name of x column.}

\item{y}{Name of y column.}

\item{value}{Optional value column for raster plots.}

\item{type}{Plot type: "line", "point", or "raster".}
}
\value{
A ggplot object.
}
\description{
A small ggplot2 wrapper for line, point, and raster/tile visualizations.
}
